max30100: Updated sample to work with d2000
Removed float print for temperature.

+++ samples/upm/max30100/src/max30100.c
@@ -72,7 +72,7 @@ int main()
         goto max30100_exit;
     }
 
-    printf("Temperature: %f C\n", temp);
+    printf("Temperature: %d C\n", (int)temp);
     printf("Version: 0x%04x\n", version);
 
     /* Set high-res (50 Hz, 16-bit) */
